Add basic lvm provider for volume_group.
This assumes that vgreduce fails when given a physical volume that is
being used by a logical volume.  If this is not the case, we need to
check pvdisplay first.  It's possible we may be able to do the pvmove
to migrate the data, but I'm not sure how that should work
(destination?) or if it's really a desirable feature.

Signed-off-by: Bruce Williams <bruce@codefluency.com>