(#9272) Refactor puppet-lvm (auto)testing
Testing needed refactoring due to recent merges and tests and
autotesting was written for rspec 1.29. This has been refreshed
for rspec 2 and autotest. There is still a little tweaking to do
to get autotest runnning out of the box.

Added .rspec and .autotest files while removing the autotesti
directory as it was needed for previous autotest. Current autotest
uses .autotest for simple configurations.

Signed-off-by: Matthaus Litteken <matthaus@puppetlabs.com>