use command 'blkid' instead of 'mount' to check the filesystem type
The options "--guess-fstype" was removed from the mount command [1].
Commands "fsck -N" and "file -sL" were also tested but their output are
not consitent between filesystems. 'blkid' finally seems to be the best candidate
to replace 'mount' and this utility is also present in the standard
package util-linux.

This patch fix the following bug https://projects.puppetlabs.com/issues/19410

[1] http://git.kernel.org/?p=utils/util-linux/util-linux.git;a=commit;h=0377ef91270d06592a0d4dd009c29e7b1ff9c9b8